fixed policy and collaborators

+++ app/controllers/collaborators_controller.rb
@@ -4,11 +4,23 @@ class CollaboratorsController < ApplicationController
     @users = User.all
     @wiki = Wiki.find(params[:wiki_id])
     @collaborator = Collaborator.new
+
+    unless current_user == @wiki.user || current_user.admin?
+      flash[:alert] = "You are not currently allowed to perform this action"
+      redirect_to root_path
+    end
+
   end
 
   def create
     @wiki = Wiki.find(params[:wiki_id])
     @collaborator = @wiki.collaborators.build(collaborator_params)
+
+    unless current_user == @wiki.user || current_user.admin?
+      flash[:alert] = "You are not currently allowed to perform this action."
+      redirect_to root_path
+    end
+
     if @collaborator.save
         flash[:notice] = "Successfully created/added to collaboratorating team"
         redirect_to @wiki
@@ -20,6 +32,12 @@ class CollaboratorsController < ApplicationController
 
   def destroy
     @collaborator = Collaborator.find(params[:id])
+    @wiki = Wiki.find(params[:wiki_id])
+
+    unless current_user.id == @wiki.user.id || current_user.admin?
+      flash[:alert] = "You are not currently allowed to perform this action."
+      redirect_to root_path
+    end
     if @collaborator.destroy
       flash[:notice] = " #{@collaborator.user.email}  was removed from this wiki."
       redirect_to :back
